+++ src/main/java/org/opentripplanner/routing/algorithm/RoutingWorker.java
@@ -1,28 +1,31 @@
 package org.opentripplanner.routing.algorithm;
 
 import java.time.Instant;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import javax.annotation.Nullable;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilterChain;
 import org.opentripplanner.routing.algorithm.mapping.RoutingRequestToFilterChainMapper;
-import org.opentripplanner.routing.algorithm.mapping.TripPlanMapper;
+import org.opentripplanner.routing.algorithm.mapping.RoutingResponseMapper;
 import org.opentripplanner.routing.algorithm.raptor.router.FilterTransitWhenDirectModeIsEmpty;
 import org.opentripplanner.routing.algorithm.raptor.router.TransitRouter;
 import org.opentripplanner.routing.algorithm.raptor.router.street.DirectFlexRouter;
 import org.opentripplanner.routing.algorithm.raptor.router.street.DirectStreetRouter;
+import org.opentripplanner.routing.algorithm.raptor.transit.mappers.DateMapper;
 import org.opentripplanner.routing.api.request.RoutingRequest;
+import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.api.response.RoutingError;
 import org.opentripplanner.routing.api.response.RoutingResponse;
-import org.opentripplanner.routing.api.response.TripSearchMetadata;
 import org.opentripplanner.routing.error.RoutingValidationException;
 import org.opentripplanner.routing.framework.DebugTimingAggregator;
 import org.opentripplanner.standalone.server.Router;
+import org.opentripplanner.transit.raptor.api.request.SearchParams;
 import org.opentripplanner.util.OTPFeature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,20 +43,24 @@ public class RoutingWorker {
 
     private final RoutingRequest request;
     private final Router router;
-    private final FilterTransitWhenDirectModeIsEmpty emptyDirectModeHandler;
+    private FilterTransitWhenDirectModeIsEmpty emptyDirectModeHandler;
+    private final ZonedDateTime searchStartTime;
+    private SearchParams raptorSearchParamsUsed = null;
     private Itinerary firstRemovedItinerary = null;
-    private Instant filterOnLatestDepartureTime = null;
-    private int searchWindowUsedInSeconds = TransitRouter.NOT_SET;
+    private boolean reverseFilteringDirection = false;
 
-    public RoutingWorker(RoutingRequest request, Router router) {
+    public RoutingWorker(Router router, RoutingRequest request, ZoneId zoneId) {
         this.request = request;
         this.router = router;
-        this.emptyDirectModeHandler = new FilterTransitWhenDirectModeIsEmpty(request.modes);
+        this.searchStartTime = DateMapper.asStartOfService(request.getDateTimeCurrentPage(), zoneId);
     }
 
     public RoutingResponse route() {
+        applyPageCursor();
+
         // If no direct mode is set, then we set one.
         // See {@link FilterTransitWhenDirectModeIsEmpty}
+        this.emptyDirectModeHandler = new FilterTransitWhenDirectModeIsEmpty(request.modes);
         request.modes.directMode = emptyDirectModeHandler.resolveDirectMode();
 
         this.debugTimingAggregator.finishedPrecalculating();
@@ -83,8 +90,9 @@ public class RoutingWorker {
         // Filter itineraries
         ItineraryListFilterChain filterChain = RoutingRequestToFilterChainMapper.createFilterChain(
             request,
-            filterOnLatestDepartureTime,
+            filterOnLatestDepartureTime(),
             emptyDirectModeHandler.removeWalkAllTheWayResults(),
+            reverseFilteringDirection,
             it -> firstRemovedItinerary = it
         );
 
@@ -99,12 +107,50 @@ public class RoutingWorker {
         // Restore original directMode.
         request.modes.directMode = emptyDirectModeHandler.originalDirectMode();
 
-        return new RoutingResponse(
-            TripPlanMapper.mapTripPlan(request, filteredItineraries),
-            createTripSearchMetadata(),
-            new ArrayList<>(routingErrors),
-            debugTimingAggregator
+        return RoutingResponseMapper.map(
+                request,
+                searchStartTime,
+                raptorSearchParamsUsed,
+                firstRemovedItinerary,
+                filteredItineraries,
+                routingErrors,
+                debugTimingAggregator,
+                reverseFilteringDirection
         );
+    }
+
+    /**
+     * Adjust the 'dateTime' if the page cursor is set to "goto next page".
+     * The date-time is used for many things, for example finding the days to search,
+     * but the transit search is using the cursor[if exist], not the date-time.
+     */
+    private void applyPageCursor() {
+        if(request.pageCursor != null) {
+            Instant dateTimeCurrentPage = request.getDateTimeCurrentPage();
+            request.setDateTime(dateTimeCurrentPage);
+            reverseFilteringDirection = request.pageCursor.reverseFilteringDirection;
+            request.modes.directMode = StreetMode.NOT_SET;
+            LOG.debug("Request dateTime={} set from pageCursor.", dateTimeCurrentPage);
+        }
+    }
+
+    /**
+     * Filter itineraries away that depart after the latest-departure-time for depart after
+     * search. These itineraries are a result of time-shifting the access leg and is needed for
+     * the raptor to prune the results. These itineraries are often not ideal, but if they
+     * pareto optimal for the "next" window, they will appear when a "next" search is performed.
+     */
+    private Instant filterOnLatestDepartureTime() {
+        if(!request.arriveBy
+            && raptorSearchParamsUsed != null
+            && raptorSearchParamsUsed.isSearchWindowSet()
+            && raptorSearchParamsUsed.isEarliestDepartureTimeSet()
+        ) {
+            int ldt = raptorSearchParamsUsed.earliestDepartureTime()
+                    + raptorSearchParamsUsed.searchWindowInSeconds();
+            return searchStartTime.plusSeconds(ldt).toInstant();
+        }
+        return null;
     }
 
     private void routeDirectStreet(
@@ -145,40 +191,18 @@ public class RoutingWorker {
     ) {
         debugTimingAggregator.startedTransitRouting();
         try {
-            var transitResults = TransitRouter.route(request, router, debugTimingAggregator);
-            filterOnLatestDepartureTime = transitResults.getFilterOnLatestDepartureTime();
-            searchWindowUsedInSeconds = transitResults.getSearchWindowUsedInSeconds();
+            var transitResults = TransitRouter.route(
+                    request,
+                    router,
+                    searchStartTime,
+                    debugTimingAggregator
+            );
+            raptorSearchParamsUsed = transitResults.getSearchParams();
             itineraries.addAll(transitResults.getItineraries());
         } catch (RoutingValidationException e) {
             routingErrors.addAll(e.getRoutingErrors());
         } finally {
             debugTimingAggregator.finishedTransitRouter();
-        }
-    }
-
-    @Nullable
-    private TripSearchMetadata createTripSearchMetadata() {
-        if(searchWindowUsedInSeconds == TransitRouter.NOT_SET) { return null; }
-
-        Instant reqTime = Instant.ofEpochSecond(request.dateTime);
-
-        if (request.arriveBy) {
-            return TripSearchMetadata.createForArriveBy(
-                reqTime,
-                searchWindowUsedInSeconds,
-                firstRemovedItinerary == null
-                    ? null
-                    : firstRemovedItinerary.endTime().toInstant()
-            );
-        }
-        else {
-            return TripSearchMetadata.createForDepartAfter(
-                reqTime,
-                searchWindowUsedInSeconds,
-                firstRemovedItinerary == null
-                    ? null
-                    : firstRemovedItinerary.startTime().toInstant()
-            );
         }
     }
 }